chore: split cli tests
