+++ tests/e2e/chrome-devtools-status.test.ts
@@ -1,0 +1,39 @@
+/**
+ * @license
+ * Copyright 2026 Google LLC
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+import assert from 'node:assert';
+import {describe, it, afterEach, beforeEach} from 'node:test';
+
+import {
+  assertDaemonIsNotRunning,
+  assertDaemonIsRunning,
+  runCli,
+} from '../utils.js';
+
+describe('chrome-devtools', () => {
+  beforeEach(async () => {
+    await runCli(['stop']);
+    await assertDaemonIsNotRunning();
+  });
+
+  afterEach(async () => {
+    await runCli(['stop']);
+    await assertDaemonIsNotRunning();
+  });
+
+  it('reports daemon status correctly', async () => {
+    await assertDaemonIsNotRunning();
+
+    const startResult = await runCli(['start']);
+    assert.strictEqual(
+      startResult.status,
+      0,
+      `start command failed: ${startResult.stderr}`,
+    );
+
+    await assertDaemonIsRunning();
+  });
+});
